further setup of examples app and the POC react component
