+++ examples/src/store.js
@@ -1,19 +1,6 @@
-import { createStore, combineReducers } from 'redux';
+import { createStore } from 'redux';
 
+import rootReducer from './reducers';
 
-const initialState = {
-  hello: 'WORLD'
-};
-
-const sampleReducer = (state = initialState, action) => {
-  const nextState = {...state};
-  switch (action.type) {
-    default:
-      break;
-  }
-  return nextState;
-};
-
-const rootReducer = combineReducers( sampleReducer );
 
 export default createStore( rootReducer );
